MessageTransform class got too complicated, let's use sprintf instead
